pyproject.toml
[tool.poetry]
name = "repl-nix-forexsignalpro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "1.28.2"
yfinance = "0.2.28"
pandas-ta = "0.3.14b0"
plotly = "5.17.0"
numpy = "1.24.3"
pandas = "2.0.3"
newsapi-python = "0.2.7"
textblob = "0.15.3"
python-telegram-bot = "20.6"
scikit-learn = "1.3.1"
scipy = "^1.14.1"
ta = "0.10.2"
statsmodels = "0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
